requirements.txt
flask==2.3.3
boto3==1.28.65
python-dotenv==1.0.0
